refactor: consolidate trivial overlay state structs + dedup migration helpers + inline 2 small overlays (closes #397)
Post-sweep YAGNI cleanup surfaced by the architectural stock-take. All
three findings cross-confirmed by ce-architecture-strategist,
ce-maintainability-reviewer, and ce-code-simplicity-reviewer agents.

A. Consolidate 11 trivial overlay state structs into src/domain/overlays.rs

Merged ActionMenuState, ContactsOverlayState, ForwardOverlayState,
PinDurationOverlayState, ThemePickerState, VerifyOverlayState,
ProfileOverlayState, GroupMenuOverlayState, KeybindingsOverlayState,
PollVoteOverlayState, and SettingsProfileOverlayState into one file.
Each was 10-30 lines of pure-data #[derive(Default)] cursor/filter
state -- one file per struct added navigation cost without payoff.
The substantive domain modules (typing, search, emoji_picker, image,
file_picker) keep their own files because they earn the separation
with real behaviour.

23 -> 13 files in src/domain/.

B. Dedup the 3 filesystem migration helpers into fs_migrate::migrate_path

config::migrate_config_dir, db::migrate_data_dir, and db::migrate_db_file
were all the same 4-line "if !new.exists() && old.exists() { mkdir
parent; rename }" shape -- two of them had identical bodies. Promoted
one shared helper at the new src/fs_migrate.rs (top-level so both the
binary tree and the lib surface used by fuzz harnesses can reach it
without exposing db's API). The Windows-rename fix from #350 now lives
in one place; future drift between the three call-sites is impossible.

Tests consolidated into fs_migrate.rs (6 cases covering: dir rename,
file rename, no-op when new exists for both kinds, no-op when old
missing, parent-creation edge case).

C. Inline 2 trivial ui/overlays/ files into their only callers

draw_customize (38 LoC, 3-row launcher) folded into ui/overlays/settings.rs.
Settings is its only caller; the file existed only because the sweep
reflexively gave every overlay its own file.

draw_delete_confirm (40 LoC, 1-line y/l/n prompt) folded into
ui/overlays/action_menu.rs. The action menu's "delete" choice is the
only entry point.

Both host-file rustdoc headers updated to mention the additional
content so the layout stays discoverable.

---

Net change: ~270 LoC removed across the tree, 13 files removed, 0
behavioural change. 528 tests pass (one config test removed because
its scenario is now covered by the consolidated migrate_path tests).
