+++ src/ui/mod.rs
@@ -19,10 +19,8 @@ use chat_pane::draw_chat_area;
 pub use links::LinkRegion;
 use links::collect_link_regions;
 use overlays::about::draw_about;
-use overlays::action_menu::draw_action_menu;
+use overlays::action_menu::{draw_action_menu, draw_delete_confirm};
 use overlays::contacts::draw_contacts;
-use overlays::customize::draw_customize;
-use overlays::delete_confirm::draw_delete_confirm;
 use overlays::emoji_picker::draw_emoji_picker;
 use overlays::file_browser::draw_file_browser;
 use overlays::forward::draw_forward;
@@ -35,7 +33,7 @@ use overlays::poll_vote::draw_poll_vote_overlay;
 use overlays::profile::draw_profile;
 use overlays::reaction_picker::draw_reaction_picker;
 use overlays::search::draw_search;
-use overlays::settings::draw_settings;
+use overlays::settings::{draw_customize, draw_settings};
 use overlays::settings_profile::draw_settings_profile_manager;
 use overlays::theme_picker::draw_theme_picker;
 use overlays::verify::draw_verify;
